fix(train): split class_decl's member helpers out for the file cap
#9886's last-wins accessor record pushed lower_decl/class_decl.rs to 2050
lines. The member-shape helpers — computed-key naming, the accessor-name
survey, and record_class_accessor itself — move to a sibling child module
beside the existing class_heritage/member_registration.

Unlike the page_meta split, this adds a child rather than renaming the
parent, so nothing keyed on the path `lower_decl/class_decl.rs` moves;
the two prose references to it elsewhere in the tree stay correct.
